Add xnat_link_data_directory variable for XNAT role
default to true
We need to distinguish between creating the data directory and linking to external storage.
If external_storage_directory is defined, and is the same path as the XNAT storage directory, then we musn't create and set permissions for this path in the XNAT role.
Otherwise the mounted storage will have the incorrect permissions.
If the external_storage_drive is the same path as xnat_web_server.storage_dir, then the following values should be set:
xnat_create_data_directory: false
xnat_link_data_directory: true